these are the paddle part of the java class=
this includes the PlayerMove class where it allows us theplayer to move our paddle sorting its position with x,y and the size with width, height. comp move automatically moves it towrads a tartget position at a fixed speed, also keeping it inside boundaries. the intersects method that i have learnt checks weather the "Puck" with center cx, cy and the radius r and is colliding with the paddle by finding the closest point on the paddle which is by calculating the distance by dx*dx + dy * dy .